backend/insert_to_entity/requirements.txt
mysqlclient >= 2.1.1
python-dotenv >= 0.21.0
pandas >= 1.5.1